requirements.txt
alpaca-py>=0.8.0
python-dotenv>=0.19.0
pandas>=1.3.0
openai>=1.0.0
numpy>=1.21.0
matplotlib>=3.4.0  # For future visualization features 